Add mpi_transpose
Adds a transpose example used via MPI. It is its own cmake project.

It will generate incrementing data as a matrix and a naive transpose is
called as a control to verify that the data was transposed correctly.

Then a transpose is done via MPI.

There is code in place to check if the results match but currently there
is an error with the MPI transpose.